Made it soo, the jump sound only plays when when is jumping and not by spacebar press. also made it so the main menu all keys besides num 1-4 do stuff. This now makes it so players can't crash the game when pressing other keys.

+++ Mario3/Final/AudioManager.cpp
@@ -79,7 +79,7 @@ void AudioManager::unmuteSounds()
 
 void AudioManager::update()
 {
-	for (int i = 0; i < mPlayingSounds.size(); ++i)
+	for (int i = 0; i < (int)mPlayingSounds.size(); ++i)
 	{
 		if (mPlayingSounds[i].currentSound.getStatus() == sf::Sound::Stopped)
 		{
